refactor: move DatabaseVacancies
